DOWNSTREAM: <carry> fix a compaction induce latency issue
The compaction behavior is changed in commit
[02635](https://github.com/etcd-io/etcd/commit/0263597ba84f65047948141696be926e53aa2429) and introduces a latency issue.
To be more speicific, the `ticker.C` acts as a fixed timer that triggers every 10ms, regardless of how long each batch of compaction takes.
This means that if a previous compaction batch takes longer than 10ms, the next batch starts immediately, making compaction a blocking operation for etcd.

To fix the issue, this commit revert the compaction to the previous behavior which ensures a 10ms delay between each batch of compaction, allowing other read and write operations to proceed smoothly.

Signed-off-by: Miancheng Lin <iml@amazon.com>